Ensure we cancel the initial panel startup if the panel was created and destroyed without ever getting control of the mainloop
This can happen if you plug in a color sensor, with the color control center
panel already open.
In this case, the g-s-d plugin creates a new instance which seemingly gets
loaded and unloaded and control is returned to to original instance.

A workaround is either to not have the control panel panel open when pushing in
the device, or to restart the control center after plugging in the hardware.