Correct sheet/minutes counts in the calibration dailog
Currently the sheet and minute counts in the calibration dialog are off.

Using the attached patch, the sheet count should be accurate again.

The minutes estimate for display calibration is harder is estimate as it
depends on the dialog and colorimeter used.

However it's probably better to estimate it to the high side as opposed to the
low side. It's better to be pleasantly surprised instead of unpleasantly
surprised.

Signed-off-by: Richard Hughes <richard@hughsie.com>