Calculate the profile whitepoint using the profile transform
For version 4 profiles, the whitepoint is set as D50 and a chromatic adaption
matrix 'chad' is used to convert this to the target whitepoint.

Using only cmsSigMediaWhitePointTag does not take chad into account, and so we
display the wrong profile temperature. Using cmsFLAGS_NOOPTIMIZE means that the
transform is cheap to set up and works for v2 and v4 profiles.