Move the profile viewer to a new repo
The profile viewer has always intended to be something useful, but we don't
want to show it in gnome-software and so it gets hidden in GNOME... which means
users don't know it's there and can only start it from the command line.

Delete the viewer here, and put it in a new repo where it can be properly
included in the desktop, and even built as a Flatpak on Flathub.